requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai
# ML/AI Libraries
torch>=2.0.0
transformers>=4.35.0
sentencepiece>=0.1.99
sacremoses>=0.1.1

# OCR
opencv-python-headless>=4.8.0
pytesseract>=0.3.10

# PDF Processing
pypdf>=4.0.0
pdfplumber>=0.10.3
pdf2image>=1.16.3
PyMuPDF>=1.23.0    
Pillow>=10.0.0
python-docx>=1.1.2

# Spreadsheets
openpyxl>=3.1.0
xlrd>=2.0.1

# Utilities
numpy>=1.24.0
sse-starlette
certifi>=2024.0.0
python-certifi-win32>=1.6.1; platform_system == "Windows"
truststore>=0.10.0